pyproject.toml
[tool.poetry]
name = "desafio_nava"
version = "0.1.0"
description = "Pipelines de dados com Dagster — Desafio Nava"
authors = ["fuyemura <fuyemura@gmail.com>"]
readme = "README.md"
packages = [{ include = "desafio_nava" }]


[tool.poetry.dependencies]
python = ">=3.11,<3.15"
pyspark = "3.5.0"
delta-spark = "3.1.0"
dagster = "^1.12.13"
dagster-webserver = "^1.12.13"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.dagster]
module_name = "desafio_nava.definitions"
[dependency-groups]
dev = [
    "ipykernel (>=7.2.0,<8.0.0)"
]
